Added Enemy Healthbar
+Most Enemies now have their own healthbar.
+Some code rewritten

+++ src Visual Studio 2013/Pew-Rework/Settings.cpp
@@ -95,7 +95,7 @@ void Settings::Update(Game &game){
 
 	//do crazy mouse stuff !!!BOOJAH!!! #21. Century
 	//play
-	if (difficulty.getGlobalBounds().intersects(sf::Rect<float>(sf::Mouse::getPosition(game.window).x, sf::Mouse::getPosition(game.window).y + 1.0f, 1.0f, 1.0f))){
+	if (difficulty.getGlobalBounds().intersects(sf::Rect<float>((float)sf::Mouse::getPosition(game.window).x, (float)sf::Mouse::getPosition(game.window).y + 1.0f, 1.0f, 1.0f))){
 		if (selection != 0){
 			selection = 0;
 			sound.PlaySound("select");
@@ -103,21 +103,21 @@ void Settings::Update(Game &game){
 	}
 
 	//again
-	if (sounds.getGlobalBounds().intersects(sf::Rect<float>(sf::Mouse::getPosition(game.window).x, sf::Mouse::getPosition(game.window).y + 1.0f, 1.0f, 1.0f))){
+	if (sounds.getGlobalBounds().intersects(sf::Rect<float>((float)sf::Mouse::getPosition(game.window).x, (float)sf::Mouse::getPosition(game.window).y + 1.0f, 1.0f, 1.0f))){
 		if (selection != 1){
 			selection = 1;
 			sound.PlaySound("select");
 		}
 	}
 	//settings
-	if (graphics.getGlobalBounds().intersects(sf::Rect<float>(sf::Mouse::getPosition(game.window).x, sf::Mouse::getPosition(game.window).y + 1.0f, 1.0f, 1.0f))){
+	if (graphics.getGlobalBounds().intersects(sf::Rect<float>((float)sf::Mouse::getPosition(game.window).x, (float)sf::Mouse::getPosition(game.window).y + 1.0f, 1.0f, 1.0f))){
 		if (selection != 2){
 			selection = 2;
 			sound.PlaySound("select");
 		}
 	}
 	//close
-	if (back.getGlobalBounds().intersects(sf::Rect<float>(sf::Mouse::getPosition(game.window).x, sf::Mouse::getPosition(game.window).y + 1.0f, 1.0f, 1.0f))){
+	if (back.getGlobalBounds().intersects(sf::Rect<float>((float)sf::Mouse::getPosition(game.window).x, (float)sf::Mouse::getPosition(game.window).y + 1.0f, 1.0f, 1.0f))){
 		if (selection != 3){
 			selection = 3;
 			sound.PlaySound("select");
